fix : 삭제 안되던 부분 온딜리트 프롭스 내려줘서 수정
